backend/requirements.txt
autopep8==2.0.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.12
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.2
